+++ src/meltano/core/runner/singer.py
@@ -22,6 +22,37 @@ class SingerPayload(IntFlag):
     STATE = 1
 
 
+class BookmarkWriter:
+    def __init__(self, job, session):
+        self.job = job
+        self.session = session
+
+    def writeline(self, line):
+        if self.job is None:
+            logging.info(
+                f"Running outside a Job context: incremental state could not be updated."
+            )
+            return
+
+        try:
+            new_state = json.loads(line)
+            job = self.job
+
+            job.payload["singer_state"] = new_state
+            job.payload_flags |= SingerPayload.STATE
+            job.save(self.session)
+
+            logging.info(f"Incremental state has been updated at {datetime.utcnow()}.")
+            logging.debug(f"Incremental state: {new_state}")
+        except Exception as err:
+            logging.warning(
+                "Received state is invalid, incremental state has not been updated"
+            )
+
+    def flush(self):
+        pass
+
+
 class SingerRunner(Runner):
     def __init__(self, elt_context: ELTContext):
         self.context = elt_context
@@ -36,25 +67,32 @@ class SingerRunner(Runner):
                 process.kill()
                 logging.error(f"{process} was killed.")
 
-    async def invoke(self, tap: PluginInvoker, target: PluginInvoker, session):
-        try:
-            # use standard pipes here because we
-            # don't need async processing between the
-            # tap and target.
-            target_in, tap_out = os.pipe()
+    async def invoke(
+        self,
+        tap: PluginInvoker,
+        target: PluginInvoker,
+        session,
+        extractor_log=None,
+        loader_log=None,
+        extractor_out=None,
+        loader_out=None,
+    ):
+        extractor_log = extractor_log or sys.stderr
+        loader_log = loader_log or sys.stderr
 
+        try:
             p_target, p_tap = None, None
-            p_target = await target.invoke_async(
-                stdin=target_in,
-                stdout=asyncio.subprocess.PIPE,  # state log
-                stderr=asyncio.subprocess.PIPE,  # Target err for logging it
-            )
-            os.close(target_in)
 
             p_tap = await tap.invoke_async(
-                stdout=tap_out, stderr=asyncio.subprocess.PIPE
+                stdout=asyncio.subprocess.PIPE,  # Singer messages
+                stderr=asyncio.subprocess.PIPE,  # Log
             )
-            os.close(tap_out)
+
+            p_target = await target.invoke_async(
+                stdin=asyncio.subprocess.PIPE,  # Singer messages
+                stdout=asyncio.subprocess.PIPE,  # Singer state
+                stderr=asyncio.subprocess.PIPE,  # Log
+            )
         except Exception as err:
             if p_tap:
                 p_tap.kill()
@@ -62,28 +100,44 @@ class SingerRunner(Runner):
                 p_target.kill()
             raise RunnerError(f"Cannot start tap or target: {err}")
 
-        # receive the target stdout and update the current job
-        # for each line
+        tap_outputs = [p_target.stdin]
+        if extractor_out:
+            tap_outputs.insert(0, extractor_out)
+
+        bookmark_writer = BookmarkWriter(self.context.job, session)
+        target_outputs = [bookmark_writer]
+        if loader_out:
+            target_outputs.insert(0, loader_out)
+
         await asyncio.wait(
             [
-                self.bookmark(p_target.stdout, session),
-                capture_subprocess_output(p_tap.stderr, sys.stderr),
-                capture_subprocess_output(p_target.stderr, sys.stderr),
-                p_target.wait(),
+                capture_subprocess_output(p_tap.stdout, *tap_outputs),
+                capture_subprocess_output(p_tap.stderr, extractor_log),
                 p_tap.wait(),
+                capture_subprocess_output(p_target.stdout, *target_outputs),
+                capture_subprocess_output(p_target.stderr, loader_log),
+                p_target.wait(),
             ],
             return_when=asyncio.FIRST_COMPLETED,
         )
 
+        # Close both sides of the tap-target pipe, so that both quit if they haven't already.
+        p_tap.stdout._transport.close()
+        p_target.stdin.close()
+
         # at this point, something already stopped, the other component
         # should die soon because of a SIGPIPE
-        target_code = await p_target.wait()
         tap_code = await p_tap.wait()
+        target_code = await p_target.wait()
 
-        if any((tap_code, target_code)):
+        if tap_code and target_code:
             raise RunnerError(
-                f"Subprocesses didn't exit cleanly: tap({tap_code}), target({target_code})"
+                f"Tap and target failed", {"extractor": tap_code, "loader": target_code}
             )
+        elif tap_code:
+            raise RunnerError(f"Tap failed", {"extractor": tap_code})
+        elif target_code:
+            raise RunnerError(f"Target failed", {"loader": target_code})
 
     def restore_bookmark(self, session, tap: PluginInvoker, full_refresh=False):
         # Delete state left over from different pipeline run for same extractor
@@ -115,42 +169,21 @@ class SingerRunner(Runner):
         else:
             logging.warning("No state was found, complete import.")
 
-    def bookmark_state(self, new_state: str, session):
-        if self.context.job is None:
-            logging.info(
-                f"Running outside a Job context: incremental state could not be updated."
-            )
-            return
-
-        try:
-            new_state = json.loads(new_state)
-            job = self.context.job
-
-            job.payload["singer_state"] = new_state
-            job.payload_flags |= SingerPayload.STATE
-            job.save(session)
-
-            logging.info(f"Incremental state has been updated at {datetime.utcnow()}.")
-            logging.debug(f"Incremental state: {new_state}")
-        except Exception as err:
-            logging.warning(
-                "Received state is invalid, incremental state has not been updated"
-            )
-
-    async def bookmark(self, target_stream, session):
-        while not target_stream.at_eof():
-            last_state = await target_stream.readline()
-            if not last_state:
-                continue
-
-            self.bookmark_state(last_state, session)
-
     def dry_run(self, tap: PluginInvoker, target: PluginInvoker):
         logging.info("Dry run:")
         logging.info(f"\textractor: {tap.plugin.name} at '{tap.exec_path()}'")
         logging.info(f"\tloader: {target.plugin.name} at '{target.exec_path()}'")
 
-    def run(self, session, dry_run=False, full_refresh=False):
+    async def run(
+        self,
+        session,
+        dry_run=False,
+        full_refresh=False,
+        extractor_log=None,
+        loader_log=None,
+        extractor_out=None,
+        loader_out=None,
+    ):
         tap = self.context.extractor_invoker()
         target = self.context.loader_invoker()
 
@@ -162,5 +195,12 @@ class SingerRunner(Runner):
 
         self.restore_bookmark(session, tap, full_refresh=full_refresh)
 
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(self.invoke(tap, target, session))
+        await self.invoke(
+            tap,
+            target,
+            session,
+            extractor_log=extractor_log,
+            loader_log=loader_log,
+            extractor_out=extractor_out,
+            loader_out=loader_out,
+        )